Moved _check_output_for_errors to cli_service
