+++ cloudshell/shell/core/command_template/command_template_service.py
@@ -34,14 +34,6 @@ class CommandTemplateService:
         CommandTemplateService._check_output_for_errors(output)
         return output
 
-    @staticmethod
-    def _check_output_for_errors(output):
-        for error_pattern in CommandTemplateService._error_list:
-            if re.search(error_pattern, output):
-                CommandTemplateService.rollback()
-                raise Exception(
-                    'Output contains error with pattern: "{0}", for output: "{1}"'.format(error_pattern, output))
-        return output
 
     @staticmethod
     def get_commands_list(command_map):
